Fixed Bugs + Fixed Welcomer

+++ Cogs/moderation.py
@@ -1,6 +1,7 @@
 # known error: 
 #   error thrown if mute role isn't there in the server
 
+from inspect import Traceback
 import discord,asyncio,json
 from discord.ext import commands
 
@@ -10,7 +11,7 @@ class Logs(commands.Cog):
     def __init__(self,bot):
         self.bot=bot
         with open("Configuration/ModConfig.json") as f: self.CONFIG = json.loads(f.read())
-        self.illegal_words=['Nigger','Nigga','N1gg3r','N1gger','Nigg3r','N1gga','N1gg@','Dick','Fuck','F U C K','f u c k','gandu','gaandu','gaamdu','fuck','nigger','nigga','n1gg3r','n1gga','n1gg@','dick']
+        self.illegal_words=['test','Nigger','Nigga','N1gg3r','N1gger','Nigg3r','N1gga','N1gg@','Dick','Fuck','F U C K','f u c k','gandu','gaandu','gaamdu','fuck','nigger','nigga','n1gg3r','n1gga','n1gg@','dick']
         
     ##########################################################################################==> This command turns the moderation system on or off, But the moderation system is turned on by default as the value of mod variable is set to true by default.
     @commands.command()
@@ -28,17 +29,27 @@ class Logs(commands.Cog):
     #############################################################################################
     @commands.Cog.listener()
     async def on_message(self,message: discord.Message) -> None:
+        if message.author.bot:
+            return
         if str(message.author.guild.id) in self.CONFIG.keys():
             if self.CONFIG[str(message.author.guild.id)]["ModEnabled"]:
                 if any(word in message.content for word in self.illegal_words):
                     user=message.author
                     await message.delete() #This command deletes the messages if it contains those words
                     await user.send('Your message was deleted due to use of profane and illegal words and you are temporarily muted for 10 minutes.')#This line sends a dm to user
-                    role=discord.utils.get(message.guild.roles,name='muted') #This command gives the user a muted role, you can change the muted role with any role you want to give but the name is case sensitive
+                    role = discord.utils.get(message.guild.roles,name='Muted') #This command gives the user a muted role, you can change the muted role with any role you want to give but the name is case sensitive
+                    
+                    if role is None:
+                        role = await message.author.guild.create_role(name="Muted")
+
+                        for channel in message.author.guild.channels:
+                            await channel.set_permissions(role, speak=False, send_messages=False, read_message_history=True, read_messages=False)
+        
+                    
                     await message.author.add_roles(role)
                     await asyncio.sleep(600.0) #this is  a timer of 10 mins, after 10 mins the role gets removed automatically.
-                    await message.author.remove_roles(role)
-                    
+                    try: await message.author.remove_roles(role)
+                    except Exception: pass
                     
         elif str(message.author.guild.id) not in self.CONFIG.keys():
             if any(word in message.content for word in self.illegal_words):
@@ -152,18 +163,16 @@ class Logs(commands.Cog):
 
     @commands.command()
     @commands.has_permissions(ban_members=True)
-    async def ban(self,ctx,user:commands.MemberConverter,*,reason=None):
+    async def ban(self,ctx,user: commands.MemberConverter,*,reason=None):
         '''This command ban a user. This command takes 2 argument out of which only one is important'''
+            
         if user == ctx.author:
             await ctx.send("You can't ban yourself!")
             return
         if user==None:
             await ctx.send('Please specify a user')
             return
-        if reason==None:
-            await user.ban(reason='Unspecified')
-        else:
-            await user.ban(reason=reason)
+        await user.ban(reason=reason)
         await ctx.send(f'{user} has been banned from your server.')
 
     @commands.command()
@@ -171,15 +180,23 @@ class Logs(commands.Cog):
     async def unban(self,ctx,*,member):
         """This command unban a banned member.You must supply the user parameter like this 'PHÄÑTÖM KÑÏGHT
 #9152'"""
-        banned_users = await ctx.guild.bans()#This command will fetch all the banned users from guild/server
-        member_name,member_disc=member.split('#')#This command will split the member name and its discriminator
-        for banned_entry in banned_users:
-            user = banned_entry.user
-            if (user.name,user.discriminator) == (member_name,member_disc):#This line checks if the given user is in banned users list if yes the next line unban them.
-                await ctx.guild.unban(user)
-                await ctx.send(member_name+'has been unbanned')
-                return
-        await ctx.send(member+'was not found')#and if the given user is not in banned users list it just send this message.
+        banned_users = await ctx.guild.bans()
+        if member.isdigit():
+            member = await self.bot.fetch_user(member)
+            if member in banned_users:
+                await ctx.send(f"{member.name} is not banned!")
+            else:
+                await ctx.guild.unban(member)
+                await ctx.send(f"Unbanned {member.name}")
+        else:
+            member_name,member_disc=member.split('#')#This command will split the member name and its discriminator
+            for banned_entry in banned_users:
+                user = banned_entry.user
+                if (user.name,user.discriminator) == (member_name,member_disc):#This line checks if the given user is in banned users list if yes the next line unban them.
+                    await ctx.guild.unban(user)
+                    await ctx.send(f'{user.name} has been unbanned')
+                    return
+            await ctx.send(member+'was not found')#and if the given user is not in banned users list it just send this message.
     
     @commands.command(aliases=['purge'])#This is a purge commands,the aliases in paranthese means that you can call this command with the folllowing names.
     @commands.has_permissions(manage_messages=True)
